Make confirmation, add alerts and links for authentication
