--- app/views/confirmations/show.html.erb
@@ -1,26 +1,0 @@
-<h1>Confirmation!!!</h1>
-
-<div class="row">
-  <div class="col-sm-2">
-  </div>
-  <div class="col-sm-8">
-    <div class="row">
-      <div class="col-sm-2">
-        <h1>Sign up</h1>
-        <div class="col-sm-12" style="background-color:#63c0de"></div>
-      </div>
-    </div>
-    <div class="row">
-      <div class="col-sm-2">
-        <%= bootstrap_form_with(url: '/user_auths/sessions', method: 'post') do |f| %>
-          <%= f.email_field :email %>
-          <%= f.password_field :password %>
-          <%= f.submit "Log in" %>
-        <% end %>
-      </div>
-
-    </div>
-  </div>
-  <div class="col-sm-2">
-  </div>
-</div>
