+++ app/views/layouts/application.html.erb
@@ -30,8 +30,16 @@
    <button class="btn btn-info navbar-btn"><%= link_to 'EN', "http://localhost:3000#{request.env['PATH_INFO'].gsub("ru", "en")}", class: "navbar-link", style: "color:#ffffff"%>/<%= link_to 'RU', "http://localhost:3000#{request.env['PATH_INFO'].gsub("en", "ru")}", class: "navbar-link", style: "color:#ffffff"%></button>
    <%= link_to('Logout', "/user_auths/sessions", method: :delete) %>
    <%= link_to('Login', "/user_auths/sessions", method: :get) %>
+   <%= link_to('Sign up', "/user_auths/registrations/", method: :get) %>
+
  </div>
 </nav>
+    <% if notice %>
+    <p class="alert alert-success"><%= notice %></p>
+  <% end %>
+  <% if alert %>
+    <p class="alert alert-danger"><%= alert %></p>
+  <% end %>
   <%= yield %>
   </body>
 </html>
