--- app/views/registrations/create.html.erb
@@ -1,2 +1,0 @@
-<h1>Hello, sweetie</h1>
-<%= params.inspect %>
