+++ app/views/registrations/show.html.erb
@@ -13,7 +13,7 @@
         <%= bootstrap_form_with model: @user, url: "/user_auths/registrations", method: "post", local: true do |f| %>
           <%= f.email_field :email %>
           <%= f.password_field :password %>
-          <%= f.submit "Sign up" %>
+          <%= f.submit "Sign up", class: "btn btn-info" %>
         <% end %>
       </div>
