+++ app/views/user_auths/show.html.erb
@@ -1,3 +1,1 @@
-<h1>User_auth</h1>
-<%= params.inspect %>
-<%= UserAuth.find_by(id: params[:id]).inspect %>
+<h1>Welcome on your page!</h1>
