+++ spec/controllers/confirmations_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe ConfirmationsController, type: :controller do
+
+end
