+++ spec/controllers/user_auths_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe UserAuthsController, type: :controller do
+
+end
